+++ src/App.css
@@ -8,60 +8,61 @@
  * Sidebar
  */
 @media (min-width: 768px) {
-    .sidebar .offcanvas-lg {
-        position: -webkit-sticky;
-        position: sticky;
-        top: 48px;
-    }
+  .sidebar .offcanvas-lg {
+    position: -webkit-sticky;
+    position: sticky;
+    top: 48px;
+  }
 
-    .navbar-search {
-        display: block;
-    }
+  .navbar-search {
+    display: block;
+  }
 }
 
 .sidebar .nav-link {
-    font-size: .875rem;
-    font-weight: 500;
+  font-size: .875rem;
+  font-weight: 500;
 }
 
 .sidebar .nav-link.active {
-    color: #2470dc;
+  color: #2470dc;
 }
 
 .sidebar-heading {
-    font-size: .75rem;
+  font-size: .75rem;
 }
 
 /*
  * Navbar
  */
 .navbar-brand {
-    padding-top: .75rem;
-    padding-bottom: .75rem;
-    font-weight: bolder;
+  padding-top: .75rem;
+  padding-bottom: .75rem;
+  font-weight: bolder;
 }
 
 .navbar .form-control {
-    padding: .75rem 1rem;
+  padding: .75rem 1rem;
 }
 
 .breadcrumb-arrow {
-    --bs-breadcrumb-divider: ">"
+  --bs-breadcrumb-divider: ">"
 }
+
 .nav-scroller {
-    position: relative;
-    z-index: 2;
-    height: 2.75rem;
-    overflow-y: hidden;
+  position: relative;
+  z-index: 2;
+  height: 2.75rem;
+  overflow-y: hidden;
 }
 
 .nav-scroller .nav {
-    display: flex;
-    flex-wrap: nowrap;
-    padding-bottom: 1rem;
-    margin-top: -1px;
-    overflow-x: auto;
-    text-align: center;
-    white-space: nowrap;
-    -webkit-overflow-scrolling: touch;
+  display: flex;
+  flex-wrap: nowrap;
+  padding-bottom: 1rem;
+  margin-top: -1px;
+  overflow-x: auto;
+  text-align: center;
+  white-space: nowrap;
+  -webkit-overflow-scrolling: touch;
 }
