+++ src/App.js
@@ -3,7 +3,7 @@ import {RouterProvider} from "react-router-dom";
 import router from "./router/Router";
 
 function App() {
-    return <RouterProvider router={router}/>;
+  return <RouterProvider router={router}/>;
 }
 
 export default App;
