+++ src/api/adminApi.js
@@ -1,17 +1,17 @@
 import axiosInstance from "./apiBase";
 
 export default {
-    getMyAdmin: async () => {
-        try {
-            const response = await axiosInstance(
-                {
-                    method: "GET",
-                    url: `/v1/my/admin`,
-                })
+  getMyAdmin: async () => {
+    try {
+      const response = await axiosInstance(
+        {
+          method: "GET",
+          url: `/v1/my/admin`,
+        })
 
-            return response.data;
-        } catch (error) {
-            return error.response;
-        }
-    },
+      return response.data;
+    } catch (error) {
+      return error.response;
+    }
+  },
 }
