+++ src/api/enumApi.js
@@ -1,17 +1,17 @@
 import axiosInstance from "./apiBase";
 
 export default {
-    getEnum: async () => {
-        try {
-            const response = await axiosInstance(
-                {
-                    method: "GET",
-                    url: `/v1/enums`,
-                }
-            );
-            return response.data;
-        } catch (error) {
-            return error.response;
+  getEnum: async () => {
+    try {
+      const response = await axiosInstance(
+        {
+          method: "GET",
+          url: `/v1/enums`,
         }
-    },
+      );
+      return response.data;
+    } catch (error) {
+      return error.response;
+    }
+  },
 }
