+++ src/index.js
@@ -8,9 +8,9 @@ import {RecoilRoot} from "recoil";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-    <RecoilRoot>
-        <App/>
-    </RecoilRoot>
+  <RecoilRoot>
+    <App/>
+  </RecoilRoot>
 );
 
 // If you want to start measuring performance in your app, pass a function
