+++ src/router/PrivateRouter.js
@@ -6,37 +6,37 @@ import {useEffect, useState} from "react";
 import Loading from "../components/common/Loading";
 
 const PrivateRouter = ({children}) => {
-    const [, setIsLoginState] = useRecoilState(LoginStatus);
-    const [isAuthenticated, setIsAuthenticated] = useState(null);
-    const [loading, setLoading] = useState(true);
+  const [, setIsLoginState] = useRecoilState(LoginStatus);
+  const [isAuthenticated, setIsAuthenticated] = useState(null);
+  const [loading, setLoading] = useState(true);
 
-    useEffect(() => {
-        const checkLogin = async () => {
-            try {
-                const response = await adminApi.getMyAdmin();
-                if (response.ok) {
-                    setIsLoginState(true);
-                    setIsAuthenticated(true);
-                } else {
-                    setIsLoginState(false);
-                    setIsAuthenticated(false);
-                }
-            } catch (e) {
-                setIsLoginState(false);
-                setIsAuthenticated(false);
-            } finally {
-                setLoading(false);
-            }
-        };
+  useEffect(() => {
+    const checkLogin = async () => {
+      try {
+        const response = await adminApi.getMyAdmin();
+        if (response.ok) {
+          setIsLoginState(true);
+          setIsAuthenticated(true);
+        } else {
+          setIsLoginState(false);
+          setIsAuthenticated(false);
+        }
+      } catch (e) {
+        setIsLoginState(false);
+        setIsAuthenticated(false);
+      } finally {
+        setLoading(false);
+      }
+    };
 
-        checkLogin();
-    }, [setIsLoginState]);
+    checkLogin();
+  }, [setIsLoginState]);
 
-    if (loading) {
-        return <Loading loading={true}/>
-    }
+  if (loading) {
+    return <Loading loading={true}/>
+  }
 
-    return isAuthenticated ? children : <Navigate to="/" replace/>;
+  return isAuthenticated ? children : <Navigate to="/" replace/>;
 };
 
 export default PrivateRouter;
