+++ src/service/LocalStorageService.js
@@ -1,11 +1,11 @@
 export const LocalStorageService = {
-    set: (key, value) => {
-        localStorage.setItem(key, value);
-    },
-    get: (key) => {
-        return localStorage.getItem(key);
-    },
-    delete: (key) => {
-        localStorage.removeItem(key);
-    },
+  set: (key, value) => {
+    localStorage.setItem(key, value);
+  },
+  get: (key) => {
+    return localStorage.getItem(key);
+  },
+  delete: (key) => {
+    localStorage.removeItem(key);
+  },
 };
