+++ src/state/LoginStatus.js
@@ -1,6 +1,6 @@
 import {atom} from 'recoil';
 
 export const LoginStatus = atom({
-    key: 'loginStatus',
-    default: false,
+  key: 'loginStatus',
+  default: false,
 });
